% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{adjacency_visualize}
\alias{adjacency_visualize}
\title{Visualize adjacency matrices.}
\usage{
adjacency_visualize(
  adjacency_matrix,
  label = NULL,
  permuted = FALSE,
  guide_break = NULL,
  n = NULL
)
}
\arguments{
\item{adjacency_matrix}{matrix. the adjacency matrix of the dynamic network at a specific time point.}

\item{label}{vector. the community label of nodes.}

\item{permuted}{logical. If TRUE, the order of nodes in the matrix is permuted according to the label of nodes, and nodes with the same label will be clustered into one group.}

\item{guide_break}{vector. If NOT NULL, this vector will be used as the aesthetic for shade of colors.}

\item{n}{the number of total nodes.}
}
\value{
a ggplot2 object.
}
\description{
Visualize dynamic networks by adjacency matrices.
}
