% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{cluster_no_selection}
\alias{cluster_no_selection}
\title{Community detection and community number selection.}
\usage{
cluster_no_selection(
  X,
  max_k,
  cluster_method = "tk_means",
  K_selection = "Silhouette",
  weight = NULL,
  parallel = TRUE
)
}
\arguments{
\item{X}{an n*d matrix. fPCA scores obtained from fPCA or fPCA_angle.}

\item{max_k}{scalar. the maximal number of communities to choose from.}

\item{cluster_method}{(optional) vector.string. if cluster_method = "k_means", use K-means method for clustering; if cluster_method = "tk_means", use trimmed K-means method for clustering;
defulat "tk_means".}

\item{K_selection}{(optional) vector.string. if K_selection = "Silhouette", use Silhouette coefficient as the criterion; if K_selection = "Pham", use the method proposed by Pham, Dimov and Nguyen for clustering. default "Silhouette".}

\item{weight}{(optional) vector. the weight for each data point when clustering.}

\item{parallel}{(optional) logical. whether to parallelize when selecting the number of communities.}
}
\value{
community membership results with the optimal number of communities.
}
\description{
A function for community detection and community number selection by performing (trimmed) K-means algorithm on fPCA scores.
}
\examples{
data("SBM_example_fASE0.rda")
result_SBM0 <- SBM_example_fASE0[[1]]
embedding_SBM0 <- fda::eval.fd(1:50, result_SBM0[[1]])
PCA_score <- fPCA(embedding_SBM0, 1:50, 2)
clus_res <- cluster_no_selection(PCA_score[[1]], 4)
}
