% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{cluster_tkmeans}
\alias{cluster_tkmeans}
\title{Perform trimmed K-means algorithm}
\usage{
cluster_tkmeans(X, k)
}
\arguments{
\item{X}{an n*d data matrix.}

\item{k}{the number of clusters K.}
}
\value{
a list containing: cluster membership and withinss.
}
\description{
Perform trimmed K-means algorithm for a dataframe.
}
