% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate cosine similarity}
\usage{
cosine_similarity(embedding_fASE, timevec)
}
\arguments{
\item{embedding_fASE}{a T*n*d array. values of the resulting latent functions obtained by fASE function at timestamp_vec.}

\item{timevec}{T vector. the observed time points of observation.}
}
\value{
an n*n matrix where each entry contains the cosine similarity of the corresponding node pair.
}
\description{
Calculate cosine similarity of latent functions for each pair of nodes in the entire time interval.
}
\examples{
data("DCBM_example_fASE0.rda")
result_DCBM0 <- DCBM_example_fASE0[[1]]
deg <- DCBM_example_fASE0[[2]]
sim <- cosine_similarity(fda::eval.fd(seq(1, 100, length = 1000), result_DCBM0[[1]]), seq(1, 100, length = 1000))
}
