% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{cosine_similarity_partial}
\alias{cosine_similarity_partial}
\title{Calculate cosine similarity for all nodes in a batch}
\usage{
cosine_similarity_partial(
  embedding_fASE,
  nodes_calc,
  timevec_list = NULL,
  timevec = NULL
)
}
\arguments{
\item{embedding_fASE}{a T*n*d array. values of the resulting latent functions obtained by fASE function at timestamp_vec.}

\item{nodes_calc}{vector. indicies of nodes included in the calculation of the cosine similarity matrix.}

\item{timevec_list}{(optional if timevec is not NULL) list. contains several index vectors of time points in timevec, each of which are used as the time period for calculation of the cosine similarity matrix.}

\item{timevec}{(optional) T vector. the observed time points of observation.}
}
\value{
a list of the same length as timevec_list, containing n*n  cosine similarity matrices for the specific time period.
}
\description{
Calculate cosine similarity of latent functions for pre-assigned nodes and all other nodes, in several pre-assigned, different intervals.
}
\examples{
data("DCBM_example_fASE0.RData")
result_DCBM0 <- DCBM_example_fASE0[[1]]
deg <- DCBM_example_fASE0[[2]]
sim <- cosine_similarity_partial(fda::eval.fd(c(seq(1, 50, length = 500), seq(51, 100, length = 500)), result_DCBM0[[1]]), 1:10, timevec_list = list(1:500, 501:1000))
}
