% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DCBM_example_fASE0}
\alias{DCBM_example_fASE0}
\title{An illustrative example where snapshots are independent DCBMs, overlaid with a core-periphery structure.}
\format{
A list containing a data frame, a vector and a list
\describe{
  \item{result_DCBM0}{a fd object containing the latent functions of accelerated fASE, including dunction fd, coefficient array, basis fd, basis splines}
  \item{deg}{The degree of nodes in the dynamic network.}
  \item{tuning_parameter0}{The hyperparameter tuning results.}
}
}
\usage{
DCBM_example_fASE0
}
\description{
A dataset containing the results of accelerated fASE, where snapshots are independent SBMs.
}
\keyword{datasets}
