% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{dd_plot}
\alias{dd_plot}
\title{draw the dd (degree-depth) plot}
\usage{
dd_plot(
  embedding_fASE,
  timestamp_vec,
  degree_vec,
  cluster = NULL,
  label = NULL,
  cluster_shape = NULL,
  label_color = NULL,
  point_size = NULL
)
}
\arguments{
\item{embedding_fASE}{a T*n*d array, containing values of the resulting latent functions obtained by fASE function at timestamp_vec.}

\item{timestamp_vec}{T vector. the observed time points of observation.}

\item{degree_vec}{n vector. the total degree of each node in the network.}

\item{cluster}{(optional) vector. the community label obtained by some clustering method.}

\item{label}{(optional) vector. the true community label.}

\item{cluster_shape}{(optional) vector. shapes for each community obtained by some clustering method.}

\item{label_color}{(optional) vector. colors for each true community.}

\item{point_size}{(optional) scalar. point size.}
}
\value{
a ggplot2 object of the dd plot.
}
\description{
Obtain the statistical depth for each node by calculating the cosine distances between latent functions, and draw the degree-depth plot.
}
\examples{
data("DCBM_example_fASE0.rda")
result_DCBM0 <- DCBM_example_fASE0[[1]]
deg <- DCBM_example_fASE0[[2]]
embedding_DCBM0 <- fda::eval.fd(1:100, result_DCBM0[[1]])
PCA_score <- fPCA_angle(embedding_DCBM0, 1:100, 20)
clus_res <- cluster_no_selection(PCA_score[[1]], 10)
C <- c(rep(1, times = 25), rep(3, times = 25), rep(5, times = 25), rep(2, times = 25), rep(4, times = 10))
ddplot <- dd_plot(embedding_DCBM0, 1:100, deg, cluster = clus_res[[2]]$cluster, label = C)
}
