% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{fPCA}
\alias{fPCA}
\title{Conduct functional PCA for latent functions without radius correction}
\usage{
fPCA(embedding_fASE, timestamp_vec, PCA_dim, active_points = NULL)
}
\arguments{
\item{embedding_fASE}{a T*n*d array. values of the resulting latent functions obtained by fASE function at timestamp_vec.}

\item{timestamp_vec}{T vector. the observed time points of observation.}

\item{PCA_dim}{scalar. the functional PCA dimension for latent functions.}

\item{active_points}{(optional) list. If the option is not NULL, it must be a list containing active time points of each node.}
}
\value{
a list of three elements: fPC scores, functional principal components, proportion of variations explained
}
\description{
Conduct functional PCA for latent functions without radius correction
}
\examples{
data("SBM_example_fASE0.rda")
result_SBM0 <- SBM_example_fASE0[[1]]
embedding_SBM0 <- fda::eval.fd(1:50, result_SBM0[[1]])
PCA_score <- fPCA(embedding_SBM0, 1:50, 2)
}
