% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{fPCA_angle}
\alias{fPCA_angle}
\title{Conduct functional PCA for latent functions with radius correction}
\usage{
fPCA_angle(embedding_fASE, timestamp_vec, PCA_dim, active_points = NULL)
}
\arguments{
\item{embedding_fASE}{an T*n*d array. values of the resulting latent functions obtained by fASE function at timestamp_vec.}

\item{timestamp_vec}{T vector. the observed time points of observation.}

\item{PCA_dim}{scalar. the functional PCA dimension for latent functions.}

\item{active_points}{(optional) list. If the option is not NULL, it must be a list containing active time points of each node.}
}
\value{
a list of three elements: fPC scores, functional principal components, proportion of variations explained
}
\description{
Conduct functional PCA for latent functions with radius correction
}
\examples{
data("DCBM_example_fASE0.rda")
result_DCBM0 <- DCBM_example_fASE0[[1]]
deg <- DCBM_example_fASE0[[2]]
embedding_DCBM0 <- fda::eval.fd(1:100, result_DCBM0[[1]])
PCA_score <- fPCA_angle(embedding_DCBM0, 1:100, 2)
}
