% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{function_sort}
\alias{function_sort}
\title{Sort latent functions into a dataframe}
\usage{
function_sort(X, timestamp_vec)
}
\arguments{
\item{X}{a T*n*d latent function value array.}

\item{timestamp_vec}{T vector. time points corresponding to the first dimension of X.}
}
\value{
a dataframe containing the latent function value for each node at each time point.
}
\description{
Transform a T*n*d latent function value array into a dataframe for the sake of plotting, without radius correction.
}
\examples{
data("SBM_example_fASE0.rda")
result_SBM0 <- SBM_example_fASE0[[1]]
embedding_SBM0 <- fda::eval.fd(1:50, result_SBM0[[1]])
node_embedding_frame0 <- function_sort(embedding_SBM0, 1:50)
}
