% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_tasks.R
\name{metric_depth}
\alias{metric_depth}
\title{Calculate the functional depth using the cosine similarity matrix}
\usage{
metric_depth(dist_mat)
}
\arguments{
\item{dist_mat}{an n*n matrix. the sine values of latent functions for each pair of nodes (1 - cosine_simlarity_matrix).}
}
\value{
vector. calculated depth for each node.
}
\description{
Calculate the functional depth of each node using the cosine similarity matrix.
}
\examples{
data("DCBM_example_fASE0.rda")
result_DCBM0 <- DCBM_example_fASE0[[1]]
deg <- DCBM_example_fASE0[[2]]
sim <- cosine_similarity(fda::eval.fd(1:100, result_DCBM0[[1]]), 1:100)
depth_stats <- metric_depth(1 - sim)
}
