% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBM_example_fASE0}
\alias{SBM_example_fASE0}
\title{An illustrative example where snapshots are independent SBMs.}
\format{
A list containing a data frame and a list:
\describe{
\item{result_SBM0}{a fd object containing the latent functions of accelerated fASE, including dunction fd, coefficient array, basis fd, basis splines}
\item{tuning_parameter0}{The hyperparameter tuning results.}
}
}
\usage{
SBM_example_fASE0
}
\description{
A dataset containing the results of accelerated fASE, where snapshots are independent SBMs.
}
\keyword{datasets}
