% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varySBM_example_fASE0}
\alias{varySBM_example_fASE0}
\title{An illustrative example where snapshots are independent DCBMs, overlaid with a core-periphery structure.}
\format{
A list containing a data frame and a vector:
\describe{
  \item{result_varySBM0}{a fd object containing the latent functions of accelerated fASE, including dunction fd, coefficient array, basis fd, basis splines}
  \item{C}{The community label of nodes in the dynamic network.}
}
}
\usage{
varySBM_example_fASE0
}
\description{
A dataset containing the results of accelerated fASE, where snapshots are independent SBMs.
}
\keyword{datasets}
